backEnd/requirements.txt
Django==2.2
Pillow==6.0.0
djangorestframework==3.9.3
